Bug 32336: (QA follow-up) Use $metadata->schema
When we worked on Koha::Biblio::Metadata, we agreed
biblio_metadata.schema would be the sourceof truth when it comes to the
record schema.

This patch acknowledges that, while retaining the fallback to the
syspref.

To test:
1. Apply this patch
2. Run:
   $ prove t/db_dependent/api/v1/biblios.t
=> SUCCESS: Tests pass, behavior is unchanged.
3. Sign off :-D

Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>
(cherry picked from commit 075ef09f48b66df7766c22e1c9ecf1d3cf7a039a)
Signed-off-by: Matt Blenkinsop <matt.blenkinsop@ptfs-europe.com>