Bug 33336: Use a dedicated column for plugin status in plugins table
In plugins table plugins/plugins-home.pl the status enable/disable is with the name, styled as a label.
With new interface ENABLED is yellow, this looks weird.
Also I bet this is bad for accessibility.

I propose to move this info to a dedicated column.
I use bootstrap classes bg-success and bg-warning like in about.pl

Test plan:
1) Install some plugins
2) Go to Administration > Manage plugins
3) Check you see table with column status
4) Disable a plugin
5) Check you see 'Disabled' in status

Signed-off-by: Laura Escamilla <laura.escamilla@bywatersolutions.com>
Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>
Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>