Bug 33346: Add Help link to Koha manual in ERM module
The 'Help' is missing in the ERM module, this patch is adding it.

Test plan:
Confirm that the Help link is displayed on the different pages of the
ERM module and that it leads to the correct page of the manual.

Signed-off-by: Pedro Amorim <pedro.amorim@ptfs-europe.com>
Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>
Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>
(cherry picked from commit 9d4c4b1cd717d314ddb6727d6354c431c8e91d15)
Signed-off-by: Matt Blenkinsop <matt.blenkinsop@ptfs-europe.com>