Bug 29311: (follow-up) Fix t/db_dependent/selenium/regressions.t
The test was checking for the value of an input field.
But with this patch set, there is no longer an input, but
a link element with the title. This adjusts the patch to look
for the correct element.

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>
(cherry picked from commit 307df8f1af4b0577faa4dbcac8072d818bc1ffe4)
Signed-off-by: Jacob O'Mara <jacobomara901@gmail.com>