Bug 33466: Link to the vendor from the ERM module
There are 3 tables in the ERM module where we display the vendor name
but without a link to the acquisition module.

This patch adds the links.

Test plan:
Create an agreement, a license and a local package.
Pick a vendor and notice that now there is a link on the list view.

Signed-off-by: David Nind <david@davidnind.com>
Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>
Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>
(cherry picked from commit f92f967626d3f3d8e1a873a5378f3fdaa0724fac)
Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>