Bug 33231: (QA follow-up) Simplify conditional for date display
This patch simply checks if we have a value for copyrighydate and
displays publicationyear if not. Even if copyrightdate is requested (MARC21)
but isn't populated, the publicatoinyear won't replace it because we
haven't transformed that field.

I think this read a bit easier, but RM can weigh in

Signed-off-by: Nick Clemens <nick@bywatersolutions.com>
Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>
(cherry picked from commit f08d5b53256f5e5db309657a7c3c48b274e379e2)
Signed-off-by: Matt Blenkinsop <matt.blenkinsop@ptfs-europe.com>