Bug 33447: (follow-up) Fix tests and make assumption explicit
The patches made an assumption that patron would always be passed. It is
within Koha, but not in the Biblios tests.

There is no scenario where we can determine pickup locations that are not in
reference to a patron (who is picking it up?) so we should always have
this parameter

Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>
(cherry picked from commit ef8adf34fde2352680934ba6a36de89e3f905a36)
Signed-off-by: Jacob O'Mara <jacobomara901@gmail.com>