Bug 33584: (follow-up) Remove DEFAULT_GENERATED from column info
MySQL 8 returns this additional information for a timestamp such
as the timestamp2 in the Commenter test.
Which results in the following alter:
    ALTER TABLE koha_myclone.database_commenter_1 MODIFY COLUMN `timestamp2` timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP DEFAULT_GENERATED COMMENT 'Column_5'
And that statement fails again on that keyword.
For now, it seems sufficient to ignore this keyword when building the alter statement.

Test plan:
Run t/db_dependent/Koha/Database/Commenter.t

Signed-off-by: Marcel de Rooy <m.de.rooy@rijksmuseum.nl>
Did test it now under MySQL 8.
Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>