Bug 22440: (follow-up) Fix existing tests
As we are introducing a FK for illrequests.biblio_id, TestBuilder is
generating a linked biblio, and so tests expecting undef are failing.

This fixes it by explicitly setting biblio_id => undef on creating the
request.

The patch also cleans the area a bit, removing a redundant and fragile
test as well.

Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>
(cherry picked from commit 5e0c6b397d169e6470f85619844f8682c62e5b07)
Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>